Refactor test logic into automation.quality.testCode.
