+++ lib/automation/quality/analyseCode.js
@@ -14,11 +14,10 @@ const analyseCode = async function ({ directory, printSuccess }) {
     throw new Error('Directory is missing.');
   }
 
-  const sourceFiles = await analyseCode.getFiles({ directory });
-
   buntstift.line();
   buntstift.info('Running code analysis...', { prefix: '▸' });
 
+  const sourceFiles = await analyseCode.getFiles({ directory });
   const doesEslintRcExist = await files.exists({ path: path.join(directory, '.eslintrc.json') });
 
   const engineConfiguration =
@@ -56,6 +55,10 @@ const analyseCode = async function ({ directory, printSuccess }) {
 };
 
 analyseCode.getFiles = async function ({ directory }) {
+  if (!directory) {
+    throw new Error('Directory is missing.');
+  }
+
   const sourceFiles = await files.get({
     directory,
     pattern: [
