+++ lib/cli/commands/release.js
@@ -11,8 +11,7 @@ const documentation = require('../../automation/documentation'),
       git = require('../../automation/git'),
       globalOptionDefinitions = require('../globalOptionDefinitions'),
       pkg = require('../../automation/pkg'),
-      quality = require('../../automation/quality'),
-      test = require('./test');
+      quality = require('../../automation/quality');
 
 const release = {
   description: 'Release a new version.',
@@ -82,7 +81,21 @@ const release = {
           throw ex;
         }
 
-        await test.run({ watch: false, type: '' });
+        try {
+          await quality.testCode({ directory });
+        } catch (ex) {
+          switch (ex.code) {
+            case 'ETESTSMISSING':
+              buntstift.warn('No tests found.');
+              break;
+            case 'ETESTSFAILED':
+              buntstift.error('Tests failed.');
+              throw ex;
+            default:
+              buntstift.error('Failed to run tests.');
+              throw ex;
+          }
+        }
 
         try {
           await pkg.checkForUnusedDependencies({ directory });
