+++ lib/cli/commands/test.js
@@ -1,43 +1,11 @@
 'use strict';
 
-const path = require('path');
-
 const buntstift = require('buntstift'),
-      flatten = require('lodash/flatten'),
-      getUsage = require('command-line-usage'),
-      oneLine = require('common-tags/lib/oneLine');
+      getUsage = require('command-line-usage');
 
-const errors = require('../../errors'),
-      execLive = require('../../shell/execLive'),
-      files = require('../../files'),
-      globalOptionDefinitions = require('../globalOptionDefinitions'),
+const globalOptionDefinitions = require('../globalOptionDefinitions'),
+      quality = require('../../automation/quality'),
       watchFilesAndExecute = require('../../watchFilesAndExecute');
-
-const runLifecycleStep = async function ({ directory, step }) {
-  if (!directory) {
-    throw new Error('Directory is misssing.');
-  }
-  if (!step) {
-    throw new Error('Step is misssing.');
-  }
-
-  const file = path.join(directory, `${step}.js`);
-
-  if (!await files.exists({ path: file })) {
-    return;
-  }
-
-  buntstift.line();
-  buntstift.info(`Running ${step} step...`, { prefix: '▸' });
-
-  /* eslint-disable global-require */
-  const stepFunction = require(file);
-  /* eslint-enable global-require */
-
-  await stepFunction();
-
-  buntstift.line();
-};
 
 const test = {
   description: 'Run tests.',
@@ -75,8 +43,6 @@ const test = {
     const directory = process.cwd(),
           { help, type, watch } = options;
 
-    const testDirectory = path.join(directory, 'test');
-
     if (help) {
       return buntstift.info(getUsage([
         { header: 'roboter test', content: this.description },
@@ -85,92 +51,19 @@ const test = {
       ]));
     }
 
-    const sourceFiles = await files.get({
-      directory,
-      pattern: [
-        '**/*.js',
-        '!node_modules/**/*.js',
-        '!coverage/**/*.js',
-        '!dist/**/*.js'
-      ]
-    });
-
     if (watch) {
       await watchFilesAndExecute({
         message: 'Running tests in watch mode...',
-        files: sourceFiles,
-        execute: async () => await test.run({ help: false, type, watch: false })
+        files: await quality.testCode.getFiles({ directory }),
+        execute: async () => await quality.testCode({ directory, type, printSuccess: true })
       });
     }
 
     try {
-      const doesTestDirectoryExist = await files.exists({ path: testDirectory });
+      await quality.testCode({ directory, type });
 
-      if (!doesTestDirectoryExist) {
-        buntstift.line();
-        buntstift.warn('No test directory found, skipping tests.');
-
-        return;
-      }
-
-      const ignoredDirectories = [ 'shared' ],
-            preferredTypes = [ 'units', 'integration', 'e2e' ];
-      let types = flatten([ type || await files.getSubDirectories({ directory: testDirectory }) ]);
-
-      types = types.filter(currentType => !ignoredDirectories.includes(currentType));
-
-      preferredTypes.reverse().forEach(preferredType => {
-        const index = types.indexOf(preferredType);
-
-        if (index !== -1) {
-          types.splice(index, 1);
-          types.unshift(preferredType);
-        }
-      });
-
-      for (let i = 0; i < types.length; i++) {
-        const currentType = types[i];
-
-        buntstift.line();
-        buntstift.info(`Running tests of type ${currentType}...`, { prefix: '▸' });
-
-        const testFiles = await files.get({
-          directory,
-          pattern: `test/${currentType}/**/*Tests.js`
-        });
-
-        const mochaOptsPath = path.join(testDirectory, currentType, 'mocha.opts');
-        const hasMochaOpts = await files.exists({ path: mochaOptsPath });
-
-        if (!testFiles || testFiles.length === 0) {
-          throw new errors.TestsMissing();
-        }
-
-        await runLifecycleStep({ directory: path.join(testDirectory, currentType), step: 'pre' });
-
-        try {
-          await execLive(oneLine`
-            npx mocha
-              --async-only
-              --bail
-              --colors
-              --exit
-              --reporter spec
-              --ui tdd
-              ${(hasMochaOpts ? `--opts ${mochaOptsPath}` : '')}
-              ${testFiles.join(' ')}
-          `, {
-            cwd: directory,
-            maxBuffer: 1000 * 1000 * 100
-          });
-        } catch (ex) {
-          throw new errors.TestsFailed();
-        } finally {
-          await runLifecycleStep({ directory: path.join(testDirectory, currentType), step: 'post' });
-        }
-
-        buntstift.success(`Tests of type ${currentType} successful.`);
-      }
+      buntstift.line();
+      buntstift.success(`Tests successful.`);
     } catch (ex) {
       switch (ex.code) {
         case 'ETESTSMISSING':
